Moved the filter input into it's own component.
