+++ src/Components/home.jsx
@@ -1,28 +1,10 @@
-import { Col, InputGroup, Row } from 'react-bootstrap';
-import PropTypes from 'prop-types';
+import { Col, Row } from 'react-bootstrap';
 import { useState } from 'react';
 import { useSelector } from 'react-redux';
 import { LinkContainer } from 'react-router-bootstrap';
 import { countries, totalConfirmed } from '../redux/data/covidData';
+import FilterInput from './FilterInput';
 import World from './world.svg';
-
-const FilterInput = (props) => {
-  const filterValues = (e) => {
-    props.setFilter(e.target.value);
-  };
-  return (
-    <InputGroup className="mb-2">
-      <InputGroup.Prepend className="input-group-prepend">
-        <i className="input-group-text fa fa-search" />
-      </InputGroup.Prepend>
-      <input type="text" onChange={filterValues} className="form-control no-shadow" />
-    </InputGroup>
-  );
-};
-
-FilterInput.propTypes = {
-  setFilter: PropTypes.func.isRequired,
-};
 
 export default function Home() {
   const [filter, setFilter] = useState('');
